Implement assisted table discovery and allow for table to be persisted
If a table name is set use it for table discovery

This allows uri's longer than 64 characters to be used. 64 characters is
the max table name length in mariadb

Also add new unit tests for table discovery

Lastly a new global flag was added delete_arrays which allows for
toggling of soft or hard deletes when `DELETE TABLE <name>` is issued.
Soft deletes are the default since the assumption is most people will
temporarily register existing arrays to run sql on