Always make sure to set all buffers for writes
We switch on the bitmap for readset for creating buffers, so we need to
make sure its set to all fields. Previously we used
`dbug_tmp_use_all_columns` but this only sets it when built in debug mode.
Instead we should use `tmp_user_all_columns`;

It is "read_set" because we use the Field->val_* accessors. So we "read"
values from the server into a "tiledb write".