Support reloading tiledb config for every query
A new configuration open is given reopen_for_every_query to disable this
new behavior. Forcing arrays to be reopened is required to allow TileDB
configuration parameters to take effect without flushing tables.
Flushing tables is a blocking operation so it might not be desired.

The downside of reopening the arrays by creating new contexts is that
any caching or in memory fragments are lost.