Correctly handle range conditions for pushdown
This combines ranges into super ranges. We also fix using epsilon for
real domains. Additional unit tests were add.

Support for OR is still not supported, this will be a follow up PR.

WHERE IN support is also not included in this PR.

both IN and OR are left for mariadb to filter