Add option for computing est table records on open
The optimizer sometimes will do a full table scan instead of index
pushdown if it things the table is too small. I've added an option for
on table loading to compute the upper bound of records in a table. This
can take several seconds on larger tables so its not ideal to all
scenarios. For now it remains disabled by default