Add support for encrypted arrays
Support for creating, reading and writing encrypted arrays is included

This includes minor modifications to create_array and discover_array in
order to allow passing the encryption string to the handler.

Due to a limitation of MariaDB if you pass an encryption key it will be
shown in the create table statement.