Add support for duplicates in sparse arrays
If an array schema has duplicates enabled we now list the dimensions as
an index instead of a primary key. We also now handle creating arrays to
match expected SQL syntax. If you set a primary key duplicates are not
allowed, else we will allow them on the newly created array.