Add support for subsetting an array on discovery
If a TileDB array has more than the 4096 max columns that MariaDB
supports we want to allow the user to issues a create table statement to
registere a subset of the dimensions/attributes. This behavior is
controlled by the new `mytile_create_allow_subset_existing_array`
session parameter. The default behavior, error if issueing a create
table with columns on an existing, is left intact as for most users they
will not want subsetting but instead likely would be trying to create a
new array.