Fix handling of string key comparison and ranges
This fixes two issues, the first being that we incorrectly were copying
the key for pushdown onto the TileDB range. We were accidentally only
considering the first character instead of all characters.

Second we were not handling partial key matches. For strings a partial
key comparison might include zero length keys, which should be
considered for any match. This adjusts to capture the comparison of zero
length string.