ref for position can't be reset with MRR
For variable length dimensions we were resetting the ref_length
and reallocating ref when the handler called for position()
based on the string size of the current coordinates. This
doesn't work for MRR cases though because the address of ref is
taken before position() is called.

The quick fix is we'll always set the ref_length to the maximum
allowed size of a row. This means we'll allocate an extra 65KB,
but in the grand scheme this is insignificant for most usages
and we only generally store a small amount of positions.