read/write_buffer_size  is now total budget
Instead of allocating based on the user set buffer size, we now let the
user set a memory budget for the total buffer size of a query. This
allows better control of the memory usage, as the user might not always
know the number of fields to be selected. This allow update to allocate
buffer based on the size in bytes of the datatype in order to more
efficiently allocate memory.