Increase default row count to 100000
Computing the actual row count is optional as it can be a costly
operation. Instead we have a hard coded value which MariaDB uses to make
determination about the query plan. The previous value of 2, was too low
and causing poor query performance as MariaDB often attempted a linear
scan instead of using the "indexes" to pushdown the ranges.

The previous value of 2 was left over from initial development and this
was suppose to have been increased along with the `records_in_range`
value.